refactor Collection (Extended Collection, deterministic ids, separation of concerns
