+++ .env.example
@@ -6,6 +6,9 @@ LOCALES=en,jp
 # Remove/comment to process all folders
 # FOLDERS_TO_PROCESS = Authorization,Users,Files,Folders
 
+# Should convert output the item creation?
+# CONVERT_LOG = true
+
 # Define the GitHub URL for each OpenAPI spec to pull process
 EN_OAS3_REPO="https://github.com/box/box-openapi.git#en"
 JP_OAS3_REPO="https://github.com/box/box-openapi.git#jp"
