+++ src/scripts/convert.js
@@ -9,6 +9,9 @@ const OPENAPI_FILENAME = 'openapi.json'
 const OPENAPI_TYPE = 'OAS3'
 const OUTPUT_FOLDER = './compiled'
 
+const FOLDERS_TO_PROCESS = process.env.FOLDERS_TO_PROCESS
+const CONVERT_LOG = process.env.CONVERT_LOG
+
 const convert = async (locale = process.argv[1]) => {
   const path = new Path(OPENAPI_TYPE, locale)
   path.translate()
@@ -16,7 +19,7 @@ const convert = async (locale = process.argv[1]) => {
   const filename = `${path.folder}/${OPENAPI_FILENAME}`
   const openapi = new OpenAPI(filename, locale)
   const openAPISpec = await openapi.process()
-  const collection = new Collection(openAPISpec, locale).process()
+  const collection = new Collection(openAPISpec, locale, FOLDERS_TO_PROCESS, CONVERT_LOG).process()
   const writer = new Writer(collection)
 
   writer.dump(OUTPUT_FOLDER, `collection.${locale}.json`)
